+++ src/lib/client.ts
@@ -16,13 +16,17 @@ type FlashContext = {
 
 const flashStores = new WeakMap<Readable<Page>, FlashContext>();
 
-const notInitialized =
-  'Flash store must be initialized with initFlash(page) before calling getFlash.';
-
+/**
+ * @deprecated Use getFlash instead.
+ */
 export function initFlash(
   page: Readable<Page>,
   options?: FlashOptions
 ): Writable<App.PageData['flash']> {
+  return _initFlash(page, options);
+}
+
+function _initFlash(page: Readable<Page>, options?: FlashOptions): Writable<App.PageData['flash']> {
   if (flashStores.has(page)) {
     // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
     return flashStores.get(page)!.store;
@@ -84,21 +88,37 @@ export function initFlash(
   return store;
 }
 
-export function getFlash(page: Readable<Page>): Writable<App.PageData['flash']> {
+/**
+ * Retrieves the flash message store for display or modification.
+ * @param page Page store, imported from `$app/stores`.
+ * @param {FlashOptions} options for the flash message. Can only be set once, usually at the highest level component where getFlash is called for the first time.
+ * @returns The flash message store.
+ */
+export function getFlash(
+  page: Readable<Page>,
+  options?: FlashOptions
+): Writable<App.PageData['flash']> {
   const context = flashStores.get(page);
-  if (!context) throw new Error(notInitialized);
+  if (!context) return _initFlash(page, options);
+  if (options)
+    throw new Error(
+      'getFlash options can only be set during the first call to getFlash. Set the options at the highest level component that calls getFlash.'
+    );
   return context.store;
 }
 
 /**
  * Updates the flash message after a fetch request.
- * @param page App page store
- * @param {Promise<void>} update A callback that is executed before the message is set, to delay the message until navigation events are completed, for example.
- * @returns {Promise<boolean>} true if a flash message existed, false if not.
+ * @param page Page store, imported from `$app/stores`.
+ * @param {Promise<void>} update A callback which is executed *before* the message is set, to delay the message until navigation events are completed, for example when using `goto`.
+ * @returns {Promise<boolean>} `true` if a flash message existed, `false` if not.
  */
 export async function updateFlash(page: Readable<Page>, update?: () => Promise<void>) {
   const store = flashStores.get(page);
-  if (!store) throw new Error(notInitialized);
+  if (!store)
+    throw new Error(
+      'Flash store must be initialized with getFlash(page) before calling updateFlash.'
+    );
 
   // Update before setting the new message, so navigation events can pass through first.
   if (update) await update();