+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Removal of flash message when navigating is now automatic - no `afterNavigate` function needed anymore.
 
+### Deprecated
+
+- `initFlash` is deprecated, use `getFlash` instead.
+- `loadFlashMessage` is deprecated, use `loadFlash` instead.
+
 ## [1.0.1] - 2023-06-24
 
 ### Fixed